uint128 math on YDai

+++ contracts/mocks/FlashMinterMock.sol
@@ -6,16 +6,16 @@ import "../interfaces/IYDai.sol";
 
 contract FlashMinterMock is IFlashMinter {
 
-    event Parameters(address user, uint256 amount, bytes data);
+    event Parameters(address user, uint128 amount, bytes data);
 
     uint256 public flashBalance;
 
-    function executeOnFlashMint(address to, uint256 yDaiAmount, bytes calldata data) external override {
+    function executeOnFlashMint(address to, uint128 yDaiAmount, bytes calldata data) external override {
         flashBalance = IYDai(msg.sender).balanceOf(address(this));
         emit Parameters(to, yDaiAmount, data);
     }
 
-    function flashMint(address yDai, uint256 amount, bytes memory data) public {
+    function flashMint(address yDai, uint128 amount, bytes memory data) public {
         IYDai(yDai).flashMint(address(this), amount, data);
     }
 }
